Refactored to use strongly typed telemetry
