+++ Cargo.toml
@@ -28,6 +28,7 @@ license = "MIT"
 clap = "4.1.8"
 criterion = { version = "0.4.0", default-features = false }
 enum-iterator = "1.2.0"
+env_logger = "0.10.0"
 expect-test = "1.4.0"
 getrandom = { version = "0.2" }
 indoc = "2.0.0"
