+++ compiler/qsc/Cargo.toml
@@ -11,6 +11,7 @@ license.workspace = true
 
 [dependencies]
 clap = { workspace = true, features = ["derive", "cargo"] }
+env_logger = { workspace = true }
 miette = { workspace = true, features = ["fancy"] }
 num-bigint = { workspace = true }
 num-complex = { workspace = true }
