+++ compiler/qsc_data_structures/Cargo.toml
@@ -10,3 +10,4 @@ repository.workspace = true
 
 [dependencies]
 miette = { workspace = true }
+serde = { workspace = true }
