+++ npm/src/log.ts
@@ -24,7 +24,9 @@ declare global {
 }
 
 export type LogLevel = "off" | "error" | "warn" | "info" | "debug" | "trace";
-export type TelemetryCollector = (event: string) => void;
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export type TelemetryEvent = { id: string; data?: any };
+export type TelemetryCollector = (event: TelemetryEvent) => void;
 
 let telemetryCollector: TelemetryCollector | null = null;
 
@@ -69,22 +71,30 @@ export const log = {
     // Utility function to ensure exhaustive type checking. See https://stackoverflow.com/a/39419171
     log.error("Exhaustive type checking didn't account for: %o", val);
   },
-  logWithLevel(level: number, ...args: any) {
+  /**
+   * @param level - A number indicating severity: 1 = Error, 2 = Warn, 3 = Info, 4 = Debug, 5 = Trace
+   * @param target - The area or component sending the messsage, e.g. "parser" (useful for filtering)
+   * @param args - The format string and args to log, e.g. ["Index of %s is %i", str, index]
+   */
+  logWithLevel(level: number, target: string, ...args: any) {
+    // Convert to a format string containing the target (if present)
+    const [, ...trailingArgs] = args; // All but first element of args
+    const outArgs = [`[%s] ${args[0]}`, target || "", ...trailingArgs];
     switch (level) {
       case 1:
-        log.error(...args);
+        log.error(...outArgs);
         break;
       case 2:
-        log.warn(...args);
+        log.warn(...outArgs);
         break;
       case 3:
-        log.info(...args);
+        log.info(...outArgs);
         break;
       case 4:
-        log.debug(...args);
+        log.debug(...outArgs);
         break;
       case 5:
-        log.trace(...args);
+        log.trace(...outArgs);
         break;
       default:
         log.error("Invalid logLevel: ", level);
@@ -93,7 +103,7 @@ export const log = {
   setTelemetryCollector(handler: TelemetryCollector) {
     telemetryCollector = handler;
   },
-  logTelemetry(event: string) {
+  logTelemetry(event: { id: string; data?: any }) {
     telemetryCollector?.(event);
   },
   isTelemetryEnabled() {
